+++ practica2/servidores/ejercicio2_tcp_ser.py
@@ -11,40 +11,55 @@ print("*****SERVIDOR*****")
 print("Nos quedamos a la espera...")
 s_cliente, addr = socketServidor.accept()
 
+sobreEscribir = 0  # usaremos para sobreescribir (o no) mas adelante
+
 eleccion = s_cliente.recv(1024)
 if eleccion.decode("utf-8") == '1':  # el cliente envia
     print("El cliente ha elegido enviar un fichero...")
     hayFichero = s_cliente.recv(1024)
-    if hayFichero.decode("utf-8") == "SI":
+    if hayFichero.decode("utf-8") == "SI":  # si el cliente tiene pdfs
         # recibimos mensaje del fichero
         FILE = s_cliente.recv(1024)
         FILE = FILE.decode("utf-8")
         if FILE == "FAIL":
             print("Algo salio mal. Cerramos conexion")
             s_cliente.close()
-            os.remove(FILE)
         else:
             print("Recibo:[" + FILE + "] del cliente con la direccion " + str(addr))
+            listaFicheros = os.listdir('.')
+            if(FILE in listaFicheros):  # si existe un fichero con el mismo nombre
+                s_cliente.send("TRUE".encode("utf-8"))
+                sobreescribir = s_cliente.recv(1024)
+                if(sobreescribir.decode("utf-8") == "N"):
+                    sobreEscribir = 1;
+                    print("El cliente no quiso sobreescribir fichero. Cerramos conexion")
+                    s_cliente.close()
+                    os.remove(FILE)
+            else:  # no existe fichero con el mismo nombre
+                s_cliente.send("FALSE".encode("utf-8"))
 
-            f = open(FILE, "wb")
-            s_cliente.send("Preparado para recibir fichero".encode("utf-8"))
-            while True:
-                chunk = s_cliente.recv(1024)
-                f.write(chunk)
-                if len(chunk) < 1024:
-                    print("Fichero recibido.")
-                    break
-            f.close()
-            s_cliente.send("Fichero recibido".encode("utf-8"))
-            s_cliente.close()
-    else:
+            # si existe fichero y el cliente quiere sobreescribir o el fichero no existe
+            if((sobreEscribir == 0 and FILE in listaFicheros) or FILE not in listaFicheros):
+                f = open(FILE, "wb")
+                s_cliente.send("Preparado para recibir fichero".encode("utf-8"))
+
+                while True:
+                    chunk = s_cliente.recv(1024)
+                    f.write(chunk)
+                    if len(chunk) < 1024:
+                        print("Fichero recibido.")
+                        break
+                f.close()
+                s_cliente.send("Fichero recibido".encode("utf-8"))
+                s_cliente.close()
+    else:  # el cliente no dispone de pdfs
         print("El cliente no tiene nada para enviar :(. Cerramos conexion")
         s_cliente.close()
 
 if eleccion.decode("utf-8") == '2':  # el cliente recibe fichero
     print("El cliente ha elegido recibir un fichero...")
     listaFicheros = os.listdir('.')
-    send_txt = json.dumps(listaFicheros)
+    send_txt = json.dumps(listaFicheros)  # asi puedo enviar una lista
     s_cliente.send(send_txt.encode("utf-8"))  # enviamos lista de ficheros al cliente
 
     fichero = s_cliente.recv(1024)  # recibimos el fichero que el cliente quiere
